requirements.txt
PyQt5
pyqt5-tools
mysql-connector-python
regex